testing/web-platform/tests/html/browsers/the-window-object/window-prototype-chain.html

<!doctype html>
<meta charset=utf-8>
<title>Prototype chain of the window object</title>
<link rel="author" title="Ms2ger" href="ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html/#window">
<link rel="help" href="http://dvcs.w3.org/hg/domcore/raw-file/tip/Overview.html#eventtarget">
<link rel="help" href="http://dev.w3.org/2006/webapi/WebIDL/#interface-prototype-object">
<link rel="help" href="http://dev.w3.org/2006/webapi/WebIDL/#named-properties-object">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
  assert_class_string(window, "Window");
}, "window object");
test(function() {
  var proto = Object.getPrototypeOf(window);
  assert_class_string(proto, "WindowPrototype");
  assert_equals(proto, Window.prototype);
}, "Window.prototype");
test(function() {
  var gsp = Object.getPrototypeOf(Object.getPrototypeOf(window));
  assert_class_string(gsp, "WindowProperties");
}, "Global scope polluter");
test(function() {
  var protoproto = Object.getPrototypeOf(Object.getPrototypeOf(Object.getPrototypeOf(window)));
  assert_class_string(protoproto, "EventTargetPrototype");
  assert_equals(protoproto, EventTarget.prototype);
}, "EventTarget.prototype");
test(function() {
  var protoprotoproto = Object.getPrototypeOf(Object.getPrototypeOf(Object.getPrototypeOf(Object.getPrototypeOf(window))));
  assert_class_string(protoprotoproto, "Object");
  assert_equals(protoprotoproto, Object.prototype);
}, "Object.prototype");
</script>
